Fix codes for ham/spam messages
Newer versions of SciPy are not deterministic about which code goes to
which cluster category. Uses argmin to get the index of the lowest
indexed unique code. From the setup of the model, this is the ham
category. Likewise, the highest indexed unique code is the spam
category. The code for the unknown value is then the one value remaining
that hasn't been used yet from 0, 1, or 2.